Split HTML sidebar renderer
