+++ Tests/QuillCodeParityTests/ParityHTMLInteractionPrimitiveGateTests.swift
@@ -208,7 +208,11 @@ final class ParityHTMLInteractionPrimitiveGateTests: QuillCodeParityTestCase {
             "WorkspaceHTMLBrowserRenderer.swift",
             "WorkspaceHTMLReviewRenderer.swift",
             "WorkspaceHTMLSecondaryPaneRenderer.swift",
+            "WorkspaceHTMLSidebarCommandRenderer.swift",
+            "WorkspaceHTMLSidebarProjectRenderer.swift",
             "WorkspaceHTMLSidebarRenderer.swift",
+            "WorkspaceHTMLSidebarSavedSearchRenderer.swift",
+            "WorkspaceHTMLSidebarThreadRenderer.swift",
             "WorkspaceHTMLTerminalRenderer.swift",
             "WorkspaceHTMLToolCardRenderer.swift",
             "WorkspaceHTMLTopBarRenderer.swift",
@@ -236,7 +240,11 @@ final class ParityHTMLInteractionPrimitiveGateTests: QuillCodeParityTestCase {
             "WorkspaceHTMLBrowserRenderer.swift",
             "WorkspaceHTMLReviewRenderer.swift",
             "WorkspaceHTMLSecondaryPaneRenderer.swift",
+            "WorkspaceHTMLSidebarCommandRenderer.swift",
+            "WorkspaceHTMLSidebarProjectRenderer.swift",
             "WorkspaceHTMLSidebarRenderer.swift",
+            "WorkspaceHTMLSidebarSavedSearchRenderer.swift",
+            "WorkspaceHTMLSidebarThreadRenderer.swift",
             "WorkspaceHTMLTerminalRenderer.swift",
             "WorkspaceHTMLToolCardRenderer.swift",
             "WorkspaceHTMLTopBarRenderer.swift",
