+++ Tests/QuillCodeParityTests/ParityHTMLRendererDelegationGateTests.swift
@@ -140,16 +140,44 @@ final class ParityHTMLRendererDelegationGateTests: QuillCodeParityTestCase {
     func testWorkspaceHTMLRendererDelegatesSidebarRendering() throws {
         let htmlText = try Self.appSourceText(named: "WorkspaceHTMLRenderer.swift")
         let sidebarText = try Self.appSourceText(named: "WorkspaceHTMLSidebarRenderer.swift")
+        let sidebarCommandText = try Self.appSourceText(named: "WorkspaceHTMLSidebarCommandRenderer.swift")
+        let sidebarProjectText = try Self.appSourceText(named: "WorkspaceHTMLSidebarProjectRenderer.swift")
+        let sidebarSavedSearchText = try Self.appSourceText(named: "WorkspaceHTMLSidebarSavedSearchRenderer.swift")
+        let sidebarThreadText = try Self.appSourceText(named: "WorkspaceHTMLSidebarThreadRenderer.swift")
 
         XCTAssertTrue(sidebarText.contains("enum WorkspaceHTMLSidebarRenderer"), "HTML sidebar rendering should live in a focused renderer.")
         XCTAssertTrue(sidebarText.contains("static func render("), "HTML sidebar rendering should expose a directly testable entry point.")
-        XCTAssertTrue(sidebarText.contains("private static func renderProjects"), "Project-list rendering should live beside sidebar HTML.")
-        XCTAssertTrue(sidebarText.contains("private static func renderThreadSections"), "Thread-section rendering should live beside sidebar HTML.")
-        XCTAssertTrue(sidebarText.contains("private static func renderBulkToolbar"), "Bulk-selection rendering should live beside sidebar HTML.")
-        XCTAssertTrue(sidebarText.contains("private static func renderFooter"), "Sidebar tool footer rendering should live beside sidebar HTML.")
-        XCTAssertTrue(sidebarText.contains("private static func renderUtilityAction"), "Individual utility command HTML should live beside grouped sidebar rendering.")
-        XCTAssertTrue(sidebarText.contains("WorkspaceHTMLPrimitives.escape"), "Sidebar renderer should reuse shared HTML escaping.")
+        XCTAssertTrue(sidebarText.contains("WorkspaceHTMLSidebarCommandRenderer.renderPrimaryActions"), "Sidebar facade should delegate primary command rendering.")
+        XCTAssertTrue(sidebarText.contains("WorkspaceHTMLSidebarThreadRenderer.render"), "Sidebar facade should delegate thread rendering.")
+        XCTAssertTrue(sidebarText.contains("WorkspaceHTMLSidebarProjectRenderer.render"), "Sidebar facade should delegate project rendering.")
+        XCTAssertTrue(sidebarText.contains("WorkspaceHTMLSidebarCommandRenderer.renderFooter"), "Sidebar facade should delegate footer command rendering.")
+        XCTAssertTrue(sidebarProjectText.contains("enum WorkspaceHTMLSidebarProjectRenderer"), "Project-list HTML should live in a focused sidebar renderer.")
+        XCTAssertTrue(sidebarProjectText.contains("private static func renderProjects"), "Project-list rendering should live beside project sidebar HTML.")
+        XCTAssertTrue(sidebarProjectText.contains("project-empty"), "Project empty-state markup should live beside project sidebar HTML.")
+        XCTAssertTrue(sidebarProjectText.contains("WorkspaceHTMLPrimitives.escape"), "Project sidebar renderer should reuse shared HTML escaping.")
+        XCTAssertTrue(sidebarThreadText.contains("enum WorkspaceHTMLSidebarThreadRenderer"), "Thread-list HTML should live in a focused sidebar renderer.")
+        XCTAssertTrue(sidebarThreadText.contains("private static func renderThreadSections"), "Thread-section rendering should live beside thread sidebar HTML.")
+        XCTAssertTrue(sidebarThreadText.contains("private static func renderBulkToolbar"), "Bulk-selection rendering should live beside thread sidebar HTML.")
+        XCTAssertTrue(sidebarThreadText.contains("private static func renderSelectionHeaderAction"), "Selection affordance rendering should stay with thread sidebar HTML.")
+        XCTAssertTrue(sidebarThreadText.contains("sidebar-empty"), "Thread empty-state markup should live beside thread sidebar HTML.")
+        XCTAssertTrue(sidebarThreadText.contains("WorkspaceHTMLPrimitives.escape"), "Thread sidebar renderer should reuse shared HTML escaping.")
+        XCTAssertTrue(sidebarSavedSearchText.contains("enum WorkspaceHTMLSidebarSavedSearchRenderer"), "Saved-search HTML should live in a focused sidebar renderer.")
+        XCTAssertTrue(sidebarSavedSearchText.contains("sidebar-saved-search-row"), "Saved-search row markup should live beside saved-search sidebar HTML.")
+        XCTAssertTrue(sidebarSavedSearchText.contains("sidebar-saved-search-create"), "Saved-search create markup should live beside saved-search sidebar HTML.")
+        XCTAssertTrue(sidebarSavedSearchText.contains("private static func renderMoveButton"), "Saved-search move buttons should live beside saved-search sidebar HTML.")
+        XCTAssertTrue(sidebarCommandText.contains("enum WorkspaceHTMLSidebarCommandRenderer"), "Sidebar command HTML should live in a focused renderer.")
+        XCTAssertTrue(sidebarCommandText.contains("static func renderFooter"), "Sidebar tool footer rendering should live beside command sidebar HTML.")
+        XCTAssertTrue(sidebarCommandText.contains("private static func renderUtilityAction"), "Individual utility command HTML should live beside grouped command rendering.")
+        XCTAssertTrue(sidebarCommandText.contains("sidebar-tools-popover"), "Sidebar footer markup should live beside command sidebar HTML.")
+        XCTAssertTrue(sidebarCommandText.contains("WorkspaceHTMLPrimitives.escape"), "Sidebar command renderer should reuse shared HTML escaping.")
         XCTAssertTrue(htmlText.contains("WorkspaceHTMLSidebarRenderer.render"), "WorkspaceHTMLRenderer should delegate sidebar rendering.")
+        XCTAssertFalse(sidebarText.contains("private static func renderProjects"), "Sidebar facade should not own project-list rendering.")
+        XCTAssertFalse(sidebarText.contains("private static func renderThreadSections"), "Sidebar facade should not own thread-section rendering.")
+        XCTAssertFalse(sidebarText.contains("private static func renderBulkToolbar"), "Sidebar facade should not own bulk-selection rendering.")
+        XCTAssertFalse(sidebarText.contains("private static func renderFooter"), "Sidebar facade should not own footer rendering.")
+        XCTAssertFalse(sidebarText.contains("private static func renderUtilityAction"), "Sidebar facade should not own individual utility command HTML.")
+        XCTAssertFalse(sidebarText.contains("sidebar-tools-popover"), "Sidebar facade should not own sidebar footer markup.")
+        XCTAssertFalse(sidebarText.contains("project-empty"), "Sidebar facade should not own project empty-state markup.")
         XCTAssertFalse(htmlText.contains("private static func renderSidebar"), "WorkspaceHTMLRenderer should not own sidebar rendering.")
         XCTAssertFalse(htmlText.contains("private static func renderSidebarPrimaryActions"), "WorkspaceHTMLRenderer should not own sidebar primary action rendering.")
         XCTAssertFalse(htmlText.contains("private static func renderSidebarSection"), "WorkspaceHTMLRenderer should not own sidebar section rendering.")
